Honor start type in .rel files when building relup files
Previously, relup file always included an application:start(Application, permanent) apply instruction for every application that appear in the UpTo/DowFrom release file, whatever their start type in the release file.
The new implementation fixes this bug by honoring the start type according to the rel(5) format. If the start type is none, no apply line is included in the relup. If the start type is load, the relup includes instruction to only load the application. Otherwise, the relup includes an instruction to start the application to the according type.
The fix is implemented by adding a new parameter to the add_application high level appup instruction. This new parameter is documented in appup(5).
Also add a test case for this fix within sasl application which simply did not have any test. The test case calls the higher level systools:make_relup function and checks properties on the output.