Fix a bug in the implementation of the pseudo-random number generator
This commit fixes an error in the mathematical formula of the Wichmann-Hill
pseudo-random number generator. In particular, the implementation used until
now produces sequences which differ from the expected ones by an extra starting
number, which is instead the very last value of the sequence.
This bug amplified the effect of extremely correlated initial numbers when
seeding different generators with very similar seed values.