Added float_to_list/2 that allows to specify formatting options
The formatting options specify the number of digits after the
decimal point when formatting floats. Additionally this new BIF
is optimized to run 5-10x faster than float_to_list/1 for floats
under 2 to the power of 52.

    float_to_list(Float, Options) -> string()

        Float   = float()
        Options = [Option]
        Option  = {precision, Precision::integer()} | compact

Returns a string which corresponds to the text representation
of Float using fixed decimal point formatting.
When precision option is specified the returned value will contain
at most Precision number of digits past the decimal point.
When compact option is provided the trailing zeros at the end of
the list are truncated.