Added supervisor:start_child/3
This new function adds a third parameter to start_child/2 which corresponds
to a child option list. At the moment only one option exists:

'delete_child_on_exit' - It tells the supervisor to delete the child spec
when the child terminates. It only has effect when the supervisor's restart
strategy is *not* simple_one_for_one.

The motivation for this feature (deleting child spec when the child terminates) is
related to the following erlang-questions and erlang-patches thread:

http://erlang.2086793.n4.nabble.com/Possible-supervisor-patch-td3092490.html

Basically, the goal is to be able to do a delete_child/2 operation followed by
a start_child/2 operation in an atomic manner.

I'm not sure however if this is the best and/or more elegant way to do it.
Either this patch or the one suggested in the previous erlang-patches thread
would allow me to accomplish my goal: atomically restart a child with different
arguments passed to its initialization function.