Remove useless casts from the emulator
Suppose an expression E has type T. Then, if E is cast to T, the cast
is useless in the sense that it does not do anything at best. At
worst, however the cast may hide an error.

This patch removes all useless casts from the ERTS emulator by means
of automatic patch generation. It uses the tool Coccinelle, a semantic
patch suite. In this case, we applied the patch:

@@
type T;
T E;
@@

-     (T)
      E